update usages affected by refactor
