+++ src/frontend/hooks/server/fields.ts
@@ -1,9 +1,8 @@
 import {useQuery} from '@tanstack/react-query';
-
-import {useProject} from './projects';
+import {useActiveProject} from '../../contexts/ActiveProjectContext';
 
 export const useFieldsQuery = () => {
-  const project = useProject();
+  const project = useActiveProject();
 
   return useQuery({
     queryKey: ['fields'],
