+++ src/frontend/hooks/server/invites.ts
@@ -3,8 +3,11 @@ import {
   useQueryClient,
   useSuspenseQuery,
 } from '@tanstack/react-query';
+
 import {useApi} from '../../contexts/ApiContext';
-import {PROJECTS_KEY, useProject, useUpdateActiveProjectId} from './projects';
+import {useActiveProject} from '../../contexts/ActiveProjectContext';
+import {usePersistedProjectId} from '../persistedState/usePersistedProjectId';
+import {ALL_PROJECTS_KEY} from './projects';
 
 export const INVITE_KEY = 'pending_invites';
 
@@ -21,23 +24,30 @@ export function usePendingInvites() {
 export function useAcceptInvite(projectId?: string) {
   const mapeoApi = useApi();
   const queryClient = useQueryClient();
-  const switchActiveProject = useUpdateActiveProjectId();
+  const switchActiveProject = usePersistedProjectId(
+    state => state.setProjectId,
+  );
 
   return useMutation({
     mutationFn: async ({inviteId}: {inviteId: string}) => {
       if (!inviteId) return;
-      mapeoApi.invite.accept({inviteId});
+      return mapeoApi.invite.accept({inviteId});
     },
     onSuccess: () => {
       // This is a workaround. There is a race condition where the project in not available when the invite is accepted. This is temporary and is currently being worked on.
       setTimeout(() => {
-        queryClient
-          .invalidateQueries({queryKey: [INVITE_KEY, PROJECTS_KEY]})
-          .then(() => {
-            if (projectId) {
-              switchActiveProject(projectId);
-            }
-          });
+        Promise.all([
+          queryClient.invalidateQueries({
+            queryKey: [INVITE_KEY],
+          }),
+          queryClient.invalidateQueries({
+            queryKey: [ALL_PROJECTS_KEY],
+          }),
+        ]).then(() => {
+          if (projectId) {
+            switchActiveProject(projectId);
+          }
+        });
       }, 5000);
     },
   });
@@ -76,7 +86,7 @@ export function useClearAllPendingInvites() {
 
 export function useSendInvite() {
   const queryClient = useQueryClient();
-  const project = useProject();
+  const project = useActiveProject();
   type InviteParams = Parameters<typeof project.$member.invite>;
   return useMutation({
     mutationFn: ({
@@ -94,7 +104,7 @@ export function useSendInvite() {
 
 export function useRequestCancelInvite() {
   const queryClient = useQueryClient();
-  const project = useProject();
+  const project = useActiveProject();
   return useMutation({
     mutationFn: (deviceId: string) =>
       project.$member.requestCancelInvite(deviceId),
