+++ src/frontend/hooks/server/media.ts
@@ -4,8 +4,8 @@ import {useMutation, useQueries} from '@tanstack/react-query';
 import {SetRequired} from 'type-fest';
 import {URL} from 'react-native-url-polyfill';
 
+import {useActiveProject} from '../../contexts/ActiveProjectContext';
 import {DraftPhoto} from '../../contexts/PhotoPromiseContext/types';
-import {useProject} from './projects';
 
 type SavablePhoto = SetRequired<
   Pick<DraftPhoto, 'originalUri' | 'previewUri' | 'thumbnailUri'>,
@@ -13,7 +13,7 @@ type SavablePhoto = SetRequired<
 >;
 
 export function useCreateBlobMutation(opts: {retry?: number} = {}) {
-  const project = useProject();
+  const project = useActiveProject();
 
   return useMutation({
     retry: opts.retry,
@@ -40,7 +40,7 @@ export function useAttachmentUrlQueries(
     Exclude<Observation['attachments'][number]['type'], 'UNRECOGNIZED'>
   >,
 ) {
-  const project = useProject();
+  const project = useActiveProject();
 
   return useQueries({
     queries: attachments.map(attachment => {
