+++ src/frontend/hooks/server/observations.ts
@@ -5,16 +5,16 @@ import {
   useQueryClient,
 } from '@tanstack/react-query';
 
-import {useProject} from './projects';
+import {useActiveProject} from '../../contexts/ActiveProjectContext';
 import {ClientGeneratedObservation} from '../../sharedTypes';
 
 export const OBSERVATION_KEY = 'observations';
 
 export function useObservations() {
-  const project = useProject();
+  const project = useActiveProject();
 
   return useSuspenseQuery({
-    queryKey: ['observations'],
+    queryKey: [OBSERVATION_KEY],
     queryFn: async () => {
       if (!project) throw new Error('Project instance does not exist');
       return project.observation.getMany();
@@ -23,7 +23,7 @@ export function useObservations() {
 }
 
 export function useObservation(observationId: string) {
-  const project = useProject();
+  const project = useActiveProject();
 
   return useSuspenseQuery({
     queryKey: [OBSERVATION_KEY, observationId],
@@ -36,7 +36,7 @@ export function useObservation(observationId: string) {
 
 export function useCreateObservation() {
   const queryClient = useQueryClient();
-  const project = useProject();
+  const project = useActiveProject();
 
   return useMutation({
     mutationFn: async ({value}: {value: ClientGeneratedObservation}) => {
@@ -55,7 +55,7 @@ export function useCreateObservation() {
 
 export function useEditObservation() {
   const queryClient = useQueryClient();
-  const project = useProject();
+  const project = useActiveProject();
 
   return useMutation({
     mutationFn: async ({
@@ -76,7 +76,7 @@ export function useEditObservation() {
 
 export function useDeleteObservation() {
   const queryClient = useQueryClient();
-  const project = useProject();
+  const project = useActiveProject();
 
   return useMutation({
     mutationFn: async ({id}: {id: string}) => {
