+++ src/frontend/hooks/server/presets.ts
@@ -4,16 +4,18 @@ import {
   useQueryClient,
   useQuery,
 } from '@tanstack/react-query';
-
-import {useProject} from './projects';
 import {PresetValue} from '@mapeo/schema';
 import {IconSize} from '../../sharedTypes';
 
+import {useActiveProject} from '../../contexts/ActiveProjectContext';
+
+export const PRESETS_KEY = 'presets';
+
 export function usePresetsQuery() {
-  const project = useProject();
+  const project = useActiveProject();
 
   return useSuspenseQuery({
-    queryKey: ['presets'],
+    queryKey: [PRESETS_KEY],
     queryFn: async () => {
       if (!project) throw new Error('Project instance does not exist');
       return await project.preset.getMany();
@@ -22,7 +24,7 @@ export function usePresetsQuery() {
 }
 
 export function usePresetsMutation() {
-  const project = useProject();
+  const project = useActiveProject();
   const queryClient = useQueryClient();
 
   return useMutation({
@@ -31,7 +33,7 @@ export function usePresetsMutation() {
       return await project.preset.create(preset);
     },
     onSuccess: () => {
-      queryClient.invalidateQueries({queryKey: ['presets']});
+      queryClient.invalidateQueries({queryKey: [PRESETS_KEY]});
     },
   });
 }
